Add files via upload
This text files contains a list of integers. This file is the copyright property and is copied from the course provided by Stanford University. I personally encourage all to join the course and get the certification as it is very useful for anyone who wants to develop his/her analytical skills for algorithm design The link for the course is given below :

Tim Roughgarden - https://www.coursera.org/learn/algorithm-design-analysis/home/welcome